Remove AnswersCaretaker - replaced with simple while loop
Change answers table schema to remove all 'answer_' prefixes from columns

Update SQL statement to preserve order for insertion

Update Answers to LinkedHashMap (preserves order)

+++ backend/src/main/java/com/jobwait/persistence/answerpersistence/IntegerAnswer.java
@@ -39,7 +39,7 @@ public class IntegerAnswer implements Answer<Integer> {
         if (this.getValue() == null) {
             statement.setNull(statementIndex, this.sqlType);
         } else {
-            statement.setInt(statementIndex, this.getValue());
+            statement.setInt(statementIndex, this.getValue().intValue());
         }
     }
